# contracts/openapi/learning.openapi.yaml
openapi: 3.0.3
info:
  title: Learning Service (Impilo Fundo) — Internal API
  description: >
    Internal API surface for **Impilo Fundo**, the **native learning
    management, certification, in-service training, pre-service training
    and CPD support capability of Impilo vNext**. Fundo is designed to
    operate as a complete modern LMS at national scale — its own
    catalogue, paths, enrolment, progress, assessments, certificates,
    learning records and analytics — without depending on any external
    LMS to function.


    External LMS platforms (Moodle, Open edX, Canvas, Chamilo, …) may be
    connected through **optional adapters** where useful, but Fundo must
    not depend on any external LMS. The existing Moodle orchestration
    endpoints documented below are one such optional adapter, preserved
    verbatim for backwards compatibility — not the Fundo runtime.


    This contract is **descriptive of the current implementation only** —
    it documents the endpoints that exist in `services/learning-service`
    today and does not assert new behaviour, new fields, or future
    surfaces. Endpoints under `/internal/v1/learning/integrations/**`
    and `/internal/v1/learning/orchestration/**` are server-to-server
    only and use shared internal keys (see `securitySchemes`), not
    end-user JWT.


    CPD authority remains with `varapi-service`. The Impilo Fundo brand
    is the product/UX name; `learning-service` is its technical home.
  version: 0.1.0
...
servers:
...
paths:
...
  /internal/v1/learning/v11/enrolments:
    post:
      tags: [FundoNativeLMS]
      summary: Create a native Fundo enrolment (idempotent)
      description: >
        Idempotent: a duplicate POST for an existing ENROLLED/IN_PROGRESS
        `(tenant, subject, course)` triple returns the existing enrolment with
        `idempotent: true`. Requires platform v1.1 headers including
        `Idempotency-Key`. Native runtime — no Moodle.
      operationId: createFundoEnrolment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [courseId, subjectType, subjectId]
              properties:
                courseId:       { type: string, format: uuid }
                subjectType:    { type: string, enum: [PROVIDER, STUDENT, STAFF, FACILITY_TEAM, OTHER] }
                subjectId:      { type: string }
                pathwayId:      { type: string, format: uuid, nullable: true }
                enrolmentType:  { type: string, enum: [SELF, ASSIGNED, COHORT, SYSTEM], default: SELF }
                assignedBy:     { type: string, nullable: true }
                dueAt:          { type: string, format: date-time, nullable: true }
      responses:
        '200':
          description: Enrolment created or returned idempotently
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '400':
          description: Invalid input
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Course not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [FundoNativeLMS]
      summary: List enrolments for a subject
      operationId: listFundoEnrolments
      parameters:
        - { name: subjectType, in: query, required: false, schema: { type: string } }
        - { name: subjectId,   in: query, required: false, schema: { type: string } }
        - { name: limit,       in: query, required: false, schema: { type: integer, default: 25, minimum: 1, maximum: 100 } }
      responses:
        '200':
          description: Enrolment list
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
  /internal/v1/learning/v11/enrolments/{enrolmentId}:
    get:
      tags: [FundoNativeLMS]
      summary: Get one enrolment
      operationId: getFundoEnrolment
      parameters:
        - { name: enrolmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Enrolment detail
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /internal/v1/learning/v11/enrolments/{enrolmentId}/cancel:
    post:
      tags: [FundoNativeLMS]
      summary: Cancel an active enrolment
      operationId: cancelFundoEnrolment
      parameters:
        - { name: enrolmentId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { reason: { type: string } }
      responses:
        '200':
          description: Cancelled (idempotent for already-cancelled/completed)
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /internal/v1/learning/v11/enrolments/{enrolmentId}/progress:
    get:
      tags: [FundoNativeLMS]
      summary: List progress rows for one enrolment
      operationId: getFundoEnrolmentProgress
      parameters:
        - { name: enrolmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Progress rows
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
  /internal/v1/learning/v11/enrolments/{enrolmentId}/certificate:
    post:
      tags: [FundoNativeLMS]
      summary: Issue a native Fundo certificate for a completed enrolment
      description: >
        Conservative issuance: only enrolments with status `COMPLETED` may
        produce certificates. Repeat calls for the same enrolment return the
        existing certificate with `idempotent: true`. This is ordinary
        certificate **metadata** — not a signed credential. Signed /
        regulator-verifiable credentials remain out of scope and the
        responsibility of `credential-verification-service` in a later phase.
      operationId: issueFundoCertificate
      parameters:
        - { name: enrolmentId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Certificate issued or returned idempotently
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '404':
          description: Enrolment not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: Enrolment not COMPLETED
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /internal/v1/learning/v11/progress:
    post:
      tags: [FundoNativeLMS]
      summary: Record native progress (lesson, module, or course scope)
      description: >
        Records a single progress write against an enrolment. Emits the
        v1.1 compliant outbox events
        `impilo.learning.progress.started.v1` and
        `impilo.learning.progress.completed.v1`. Setting both `moduleId`
        and `lessonId` to null treats the write as a course-aggregate
        update; reaching 100 % on the aggregate transitions the enrolment
        to COMPLETED and emits `impilo.learning.course.completed.v1`.
      operationId: recordFundoProgress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrolmentId, progressPercent]
              properties:
                enrolmentId:       { type: string, format: uuid }
                moduleId:          { type: string, format: uuid, nullable: true }
                lessonId:          { type: string, format: uuid, nullable: true }
                progressPercent:   { type: integer, minimum: 0, maximum: 100 }
                score:             { type: integer, nullable: true }
      responses:
        '200':
          description: Progress recorded
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
        '400':
          description: Invalid input
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '404':
          description: Enrolment / lesson / module not found
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: Enrolment not active
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    get:
      tags: [FundoNativeLMS]
      summary: List progress rows for a subject
      operationId: listFundoProgress
      parameters:
        - { name: subjectType, in: query, required: false, schema: { type: string } }
        - { name: subjectId,   in: query, required: false, schema: { type: string } }
      responses:
        '200':
          description: Progress rows
          content: { application/json: { schema: { $ref: '#/components/schemas/ApiEnvelope_Map' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Used by the default security chain when `learning.security.oauth2-enabled=true`.
        Compose dev environments run with this disabled (see
        `ops/runtime/docker-compose.operations.yml`); Helm defaults to true.
    integrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Internal-Key
      description: >
        Shared internal key for trusted ingress from `varapi-service` after a
        Fundo webhook creates a governed CPD candidate.
    orchestrationKey:
      type: apiKey
      in: header
      name: X-Impilo-Learning-Orchestration-Key
      description: >
        Shared internal key for server-to-server orchestration jobs (Tshepo /
        Varapi schedulers, Moodle Web Service relays).
  schemas:
...
    ApiEnvelope_Map:
      type: object
      properties:
        data:
          type: object
          additionalProperties: true
      required: [data]
...
    Error:
      type: object
      properties:
        error: { type: string }

    # ── Request bodies ────────────────────────────────────────────────────
...